Add files via upload
This is a web scraping project where I took data from the IMBD Top 100. In this project, suggestions for films with the most watched genres will be displayed, as well as a data visualization in ggplot, which will provide insight into the biggest-grossing films, the most watched genres, and the best ratings in any given year.